Refactor PanelMixin class into its own file
